fix metadata and transcript loading on Vercel bc of ref not loading in time
